+++ Cargo.toml
@@ -15,4 +15,3 @@ tempfile = "3.14.0"
 statrs = "0.17.1"
 serde = "1.0.217"
 serde_derive = "1.0.217"
-ordered-float = "4.6.0"
